Migrate to react router 7
