+++ frontend/.gitignore
@@ -7,3 +7,4 @@ node_modules/
 /playwright-report/
 /blob-report/
 /playwright/.cache/
+.react-router/
