+++ frontend/package.json
@@ -9,11 +9,10 @@
   "dependencies": {
     "@monaco-editor/react": "^4.6.0",
     "@nextui-org/react": "^2.4.8",
+    "@react-router/node": "^7.0.1",
+    "@react-router/serve": "^7.0.1",
     "@react-types/shared": "^3.25.0",
     "@reduxjs/toolkit": "^2.3.0",
-    "@remix-run/node": "^2.11.2",
-    "@remix-run/react": "^2.11.2",
-    "@remix-run/serve": "^2.11.2",
     "@tanstack/react-query": "^5.60.5",
     "@vitejs/plugin-react": "^4.3.2",
     "@xterm/addon-fit": "^0.10.0",
@@ -35,7 +34,7 @@
     "react-icons": "^5.3.0",
     "react-markdown": "^9.0.1",
     "react-redux": "^9.1.2",
-    "react-router-dom": "^6.26.1",
+    "react-router": "^7.0.1",
     "react-syntax-highlighter": "^15.6.1",
     "react-textarea-autosize": "^8.5.4",
     "remark-gfm": "^4.0.0",
@@ -47,9 +46,9 @@
     "ws": "^8.18.0"
   },
   "scripts": {
-    "dev": "npm run make-i18n && cross-env VITE_MOCK_API=false remix vite:dev",
-    "dev:mock": "npm run make-i18n && cross-env VITE_MOCK_API=true remix vite:dev",
-    "build": "npm run make-i18n && tsc && remix vite:build",
+    "dev": "npm run make-i18n && cross-env VITE_MOCK_API=false react-router dev",
+    "dev:mock": "npm run make-i18n && cross-env VITE_MOCK_API=true react-router dev",
+    "build": "npm run make-i18n && tsc && react-router build",
     "start": "npx sirv-cli build/ --single",
     "test": "vitest run",
     "test:e2e": "playwright test",
@@ -60,7 +59,8 @@
     "prelint": "npm run make-i18n",
     "lint": "eslint src --ext .ts,.tsx,.js && prettier --check src/**/*.{ts,tsx}",
     "lint:fix": "eslint src --ext .ts,.tsx,.js --fix && prettier --write src/**/*.{ts,tsx}",
-    "prepare": "cd .. && husky frontend/.husky"
+    "prepare": "cd .. && husky frontend/.husky",
+    "typecheck": "react-router typegen && tsc"
   },
   "husky": {
     "hooks": {
@@ -75,8 +75,7 @@
   },
   "devDependencies": {
     "@playwright/test": "^1.48.2",
-    "@remix-run/dev": "^2.11.2",
-    "@remix-run/testing": "^2.11.2",
+    "@react-router/dev": "^7.0.1",
     "@tailwindcss/typography": "^0.5.15",
     "@tanstack/eslint-plugin-query": "^5.60.1",
     "@testing-library/jest-dom": "^6.6.1",
