+++ frontend/src/root.tsx
@@ -1,11 +1,4 @@
-import {
-  Links,
-  Meta,
-  MetaFunction,
-  Outlet,
-  Scripts,
-  ScrollRestoration,
-} from "@remix-run/react";
+import { Links, Meta, MetaFunction, Outlet, Scripts, ScrollRestoration } from "react-router";
 import "./tailwind.css";
 import "./index.css";
 import React from "react";
