+++ frontend/vite.config.ts
@@ -1,10 +1,9 @@
-/* eslint-disable import/no-extraneous-dependencies */
 /// <reference types="vitest" />
 /// <reference types="vite-plugin-svgr/client" />
 import { defineConfig, loadEnv } from "vite";
 import viteTsconfigPaths from "vite-tsconfig-paths";
 import svgr from "vite-plugin-svgr";
-import { vitePlugin as remix } from "@remix-run/dev";
+import { reactRouter } from "@react-router/dev/vite";
 import { configDefaults } from "vitest/config";
 
 export default defineConfig(({ mode }) => {
@@ -24,51 +23,9 @@ export default defineConfig(({ mode }) => {
   const WS_URL = `${WS_PROTOCOL}://${VITE_BACKEND_HOST}/`;
   const FE_PORT = Number.parseInt(VITE_FRONTEND_PORT, 10);
 
-  /**
-   * This script is used to unpack the client directory from the frontend build directory.
-   * Remix SPA mode builds the client directory into the build directory. This function
-   * moves the contents of the client directory to the build directory and then removes the
-   * client directory.
-   *
-   * This script is used in the buildEnd function of the Vite config.
-   */
-  const unpackClientDirectory = async () => {
-    const fs = await import("fs");
-    const path = await import("path");
-
-    const buildDir = path.resolve(__dirname, "build");
-    const clientDir = path.resolve(buildDir, "client");
-
-    const files = await fs.promises.readdir(clientDir);
-    await Promise.all(
-      files.map((file) =>
-        fs.promises.rename(
-          path.resolve(clientDir, file),
-          path.resolve(buildDir, file),
-        ),
-      ),
-    );
-
-    await fs.promises.rmdir(clientDir);
-  };
-
   return {
     plugins: [
-      !process.env.VITEST &&
-        remix({
-          future: {
-            v3_fetcherPersist: true,
-            v3_relativeSplatPath: true,
-            v3_throwAbortReason: true,
-            v3_lazyRouteDiscovery: true,
-            v3_singleFetch: true,
-            v3_routeConfig: true,
-            unstable_optimizeDeps: true,
-          },
-          appDirectory: "src",
-          buildEnd: unpackClientDirectory,
-          ssr: false,
-        }),
+      !process.env.VITEST && reactRouter(),
       viteTsconfigPaths(),
       svgr(),
     ],
